x64 Phase 2b (final) + Phase 2c: WARNINGS=0 across all four configs
Phase 2b chunk 4 — the long tail of remaining narrowings:

  app/cedtAppFile.cpp           DoModal() returns INT_PTR
  app/cedtAppHndr.cpp           DoModal() returns INT_PTR
  app/cedtAppSearch.cpp         strlen at folder-recursion entry
  dialogs/FindDialog.cpp        m_lstFindString.GetCount()
  dialogs/FindInFilesDialog.cpp m_lstFindString.GetCount() ×2
                                + arrFilterExtensions.GetSize()
  dialogs/OpenRemoteDialog.cpp  m_arrFilterDescription.GetSize()
                                + arrFileInfo.GetSize()
  dialogs/ReplaceDialog.cpp     m_lstFindString.GetCount() ×2
  dialogs/preferences/
    prefdialog.cpp              NM_TREEVIEW::itemNew.lParam (LPARAM)
    PrefDialogColors.cpp        strlen + DoModal() ×3
    PrefDialogCommands.cpp      strlen at every magic-header read/write
    PrefDialogMacros.cpp        same
  doc/cedtDocAnal.cpp           (handled via signature change below)
  doc/cedtDocDictionary.cpp     file.GetLength() narrows to (UINT)
  doc/cedtDocEdit.cpp           rBlock.GetCount() + strlen
  doc/cedtDocHndr.cpp           CFileStatus::m_size narrows to (long)
                                (matches DlgPropertiesFile::m_nFileSize)
  frame/MainFrm.cpp             m_lstChildFrame.GetCount()
  frame/StatusBarEx.cpp         strlen for status-pane width measurement
  panels/FileWndDirectory.cpp   m_arrFilterDescription.GetSize()
  util/CmdLine.cpp              m_arrArguments.GetSize() ×2 + strlen
  util/HyperLink.cpp            ReportError takes int, result is INT_PTR
  util/registry.cpp             RegSetValueEx cbData is DWORD, not size_t
  util/SortStringArray.cpp      GetUpperBound() returns INT_PTR
  view/cedtViewCommand.cpp      Left()/Mid() take int, not __int64 diff
  view/cedtViewEdit.cpp         strlen
  view/cedtViewEditAdv.cpp      rBlock.GetCount() + strlen
  view/cedtViewEditCompose.cpp  strlen ×2
  view/cedtViewMacro.cpp        arrPathName.GetSize()
  view/cedtViewMap.cpp          GetCount() in y-pixel span calc
  view/cedtViewMisc.cpp         GlobalSize() returns SIZE_T
  view/cedtViewPrint.cpp        m_clsFormatedPrintText.GetCount()

  core/cedtElement.h/cpp        CKeywords::LookupTable and
                                CDictionary::LookupTable signatures
                                widened from (SHORT siLength) to
                                (INT_PTR siLengthPtr) with internal
                                SHORT narrow after the MAX_WORD_LENGTH
                                guard — same trick we used for
                                _WordFound / _WordFoundExtended. This
                                quietly removed the 4 remaining
                                __int64-to-SHORT warnings at the
                                cedtDocAnal.cpp LookupTable callsites
                                without forcing a cast at every caller.

Phase 2c — non-narrowing leftovers:

  C4996 deprecated POSIX names — straight underscore prefix:
    doc/cedtDocEditAdv.cpp:200,202   strnicmp  → _strnicmp
    panels/FileWndDirectory.cpp:119,424  strnicmp → _strnicmp
    util/evaluate.cpp:343,361        strlwr    → _strlwr
    util/PathName.cpp:57             stricmp   → _stricmp

  C4996 Enable3dControls — removed the dead pre-VS2002 call entirely
                           (was inside #ifdef _AFXDLL); the symmetric
                           Enable3dControlsStatic() is gone from MFC
                           too, so the whole #ifdef block is dropped.

  C4996 GetVersionEx — kept (XP-era code-path PACKVERSION lookup that
                       still happens to work); silenced with a
                       localized #pragma warning(push)/(disable:4996).

  C5033 'register' — removed in C++17; deleted the storage class on
                     two locals in encode.cpp.

  C4838 UINT → CLIPFORMAT brace-init narrowing — explicit (CLIPFORMAT)
                     cast at FORMATETC initialization in FileTab.cpp
                     and FileWndProject.cpp.

  C4091 'CRect ;' stray declaration with no variable name in
                     XPTabCtrl.cpp:105 — just removed the dead line.

  C4090 SHBrowseForFolder takes __unaligned LPBROWSEINFOW in modern
                     SDK; matched with two casts at the single
                     callsite in FolderDialog.cpp.

Final state — all four x64 configurations build clean:

                Errors  Warnings
  Debug-KR        0       0
  Debug-US        0       0
  Release-KR      0       0
  Release-US      0       0

  cedt_tests x64 still passes 60/60 in Debug.

Binary sizes vs the original Win32 build:

  Release-KR  1.54 MB (Win32)  →  1.61 MB (x64)
  Release-US  1.56 MB (Win32)  →  1.63 MB (x64)
  Debug-KR    4.28 MB (Win32)  →  4.49 MB (x64)
  Debug-US    4.30 MB (Win32)  →  4.51 MB (x64)

Cumulative warning cleanup since the start of Phase 2:
  start : 756
  end   :   0   (100% removed)

+++ src/app/cedtAppFile.cpp
@@ -67,7 +67,7 @@ void CCedtApp::OnFileOpenRemote()
 		dlg.SetFtpAccounts( MAX_FTP_ACCOUNT, m_clsFtpAccounts );
 		dlg.SetCurrentFtpAccount( nFtpAccount );
 		dlg.SetCurrentFilterIndex( GetFilterIndexDialog() );
-		INT nResponse = dlg.DoModal();
+		INT nResponse = (INT)dlg.DoModal();
 
 		dlg.GetFtpAccounts( MAX_FTP_ACCOUNT, m_clsFtpAccounts );
 		SaveFtpAccountInfo(m_szAppDataDirectory + "\\cedt.ftp");
@@ -243,7 +243,7 @@ BOOL CCedtApp::PromptSaveRemoteFileName(INT & nFtpAccount, CString & szPathName)
 	dlg.SetFtpAccounts( MAX_FTP_ACCOUNT, m_clsFtpAccounts );
 	dlg.SetCurrentFtpAccount( nFtpAccount );
 	dlg.SetCurrentFilterIndex( GetFilterIndexDialog() );
-	INT nResponse = dlg.DoModal();
+	INT nResponse = (INT)dlg.DoModal();
 
 	dlg.GetFtpAccounts( MAX_FTP_ACCOUNT, m_clsFtpAccounts );
 	SaveFtpAccountInfo(m_szAppDataDirectory + "\\cedt.ftp");
