+++ src/util/HyperLink.cpp
@@ -551,7 +551,7 @@ void CHyperLink::FollowLink()
 	INT_PTR result = (INT_PTR) GotoURL(m_strURL, SW_SHOW);
     if (result <= (INT_PTR)HINSTANCE_ERROR) {
         MessageBeep(MB_ICONEXCLAMATION);	// Unable to follow link
-        ReportError(result);
+        ReportError((int)result);
     } else {
 		// Mark link as visited and repaint window
 		m_bVisited = TRUE;
